commit nov 7

+++ Nov5_Class/operators_12Nov/task1.py
@@ -1,5 +1,5 @@
 #/Write a program to take a user age and let him know if he can go the club.  21
-age=float(input("Enter your age:"))
+age=int(input("Enter your age:"))
 if(age>=21):
     print("Allowed to club")
 else:
